--- a/rspace/src/main/scala/coop/rchain/rspace/InMemoryStore.scala
+++ b/rspace/src/main/scala/coop/rchain/rspace/InMemoryStore.scala
@@ -225,9 +225,9 @@ class InMemoryStore[T, C, P, A, K](
 
   protected def processTrieUpdate(update: TrieUpdate[C, P, A, K]): Unit =
     update match {
-      case TrieUpdate(_, Insert, channelsHash, gnat) =>
+      case TrieUpdate(Insert, channelsHash, gnat) =>
         history.insert(trieStore, trieBranch, channelsHash, canonicalize(gnat))
-      case TrieUpdate(_, Delete, channelsHash, gnat) =>
+      case TrieUpdate(Delete, channelsHash, gnat) =>
         history.delete(trieStore, trieBranch, channelsHash, canonicalize(gnat))
     }