--- a/rspace/src/main/scala/coop/rchain/rspace/LockFreeInMemoryStore.scala
+++ b/rspace/src/main/scala/coop/rchain/rspace/LockFreeInMemoryStore.scala
@@ -216,9 +216,9 @@ class LockFreeInMemoryStore[T, C, P, A, K](
 
   protected def processTrieUpdate(update: TrieUpdate[C, P, A, K]): Unit =
     update match {
-      case TrieUpdate(_, Insert, channelsHash, gnat) =>
+      case TrieUpdate(Insert, channelsHash, gnat) =>
         history.insert(trieStore, trieBranch, channelsHash, canonicalize(gnat))
-      case TrieUpdate(_, Delete, channelsHash, gnat) =>
+      case TrieUpdate(Delete, channelsHash, gnat) =>
         history.delete(trieStore, trieBranch, channelsHash, canonicalize(gnat))
     }