--- a/rspace/src/main/scala/coop/rchain/rspace/RSpaceOps.scala
+++ b/rspace/src/main/scala/coop/rchain/rspace/RSpaceOps.scala
@@ -115,7 +115,7 @@ abstract class RSpaceOps[F[_], C, P, E, A, R, K](val store: IStore[C, P, A, K], 
           store.trieStore.validateAndPutRoot(trieTxn, store.trieBranch, root)
           val leaves = store.trieStore.getLeaves(trieTxn, root)
           eventLog.update(const(Seq.empty))
-          store.clearTrieUpdates()
+          store.getAndClearTrieUpdates()
           store.clear(txn)
           restoreInstalls(txn)
           store.bulkInsert(txn, leaves.map { case Leaf(k, v) => (k, v) })